requirements.txt
Django==5.0
Pillow==10.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-jazzmin==2.6.0
psycopg==3.1.14
django-simple-history==3.4.0
django-filter==23.5
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
drf-yasg==1.21.7
swagger-spec-validator==3.0.3
psycopg2-binary==2.9.9
